+++ packages/ui-components/src/common/utils/charts.ts
@@ -1,0 +1,8 @@
+export interface BaseChartProps {
+  width: number;
+  height: number;
+  marginTop?: number;
+  marginBottom?: number;
+  marginLeft?: number;
+  marginRight?: number;
+}
