--- packages/ui-components/src/components/MetricLineChart/interfaces.ts
@@ -1,8 +1,0 @@
-import { Metric } from "@actnowcoalition/metrics";
-import { Region } from "@actnowcoalition/regions";
-import { BaseChartProps } from "../TimeseriesLineChart";
-
-export interface MetricLineChartProps extends BaseChartProps {
-  metric: Metric | string;
-  region: Region;
-}
