+++ packages/ui-components/src/components/MetricSeriesChart/index.ts
@@ -1,2 +1,1 @@
-export * from "./interfaces";
 export * from "./MetricSeriesChart";
