+++ packages/ui-components/src/components/SeriesChart/index.ts
@@ -1,0 +1,1 @@
+export * from "./SeriesChart";
