+++ packages/ui-components/src/components/TimeseriesLineChart/index.ts
@@ -1,2 +1,1 @@
-export * from "./interfaces";
 export * from "./TimeseriesLineChart";
